refactors api/cli
