+++ conan/cli/commands/install.py
@@ -4,7 +4,8 @@ import os
 from conan.api.output import ConanOutput, cli_out_write
 from conan.cli.command import conan_command, Extender, COMMAND_GROUPS
 from conan.cli.commands import make_abs_path
-from conan.cli.common import get_profiles_from_args, get_lockfile, get_multiple_remotes, scope_options, save_lockfile_out
+from conan.cli.common import get_profiles_from_args, get_lockfile, scope_options, save_lockfile_out
+from conan.api.subapi.remotes import get_multiple_remotes
 from conan.cli.args import add_lockfile_args, _add_common_install_arguments, add_reference_args, \
     _help_build_policies
 from conan.cli.printers.graph import print_graph_basic, print_graph_packages
