+++ conan/cli/commands/test.py
@@ -4,7 +4,8 @@ from conan.api.output import ConanOutput
 from conan.cli.command import conan_command, COMMAND_GROUPS, OnceArgument
 from conan.cli.commands.create import test_package, _check_tested_reference_matches
 from conan.cli.commands.install import _get_conanfile_path
-from conan.cli.common import get_lockfile, get_profiles_from_args, get_multiple_remotes, save_lockfile_out
+from conan.cli.common import get_lockfile, get_profiles_from_args, save_lockfile_out
+from conan.api.subapi.remotes import get_multiple_remotes
 from conan.cli.args import add_lockfile_args, _add_common_install_arguments
 from conan.cli.printers.graph import print_graph_basic, print_graph_packages
 from conans.model.recipe_ref import RecipeReference
